+++ jooby/src/main/java/io/jooby/problem/HttpProblemMappable.java
@@ -1,0 +1,6 @@
+package io.jooby.problem;
+
+public interface HttpProblemMappable {
+
+  HttpProblem toHttpProblem();
+}
